Lighting, Spawns and Sprites
Adjusted some Lighting to the Game to a closer look to the final game.

Safe Spawning is now implemented, to prevent the player from splitting dangerously.

Some environment Items now rotate with the player. Spikes rotate, but not exactly how id like to, will figure out later...

Fixed weird jumping bug in the z direction because i used a Vector2 instead of a Vector3